Fix test case
Now that we use fugitive's function, the remote url wasn't working
correctly in tests.
Presumably before, because we were manually running the git commands but
now fugitive determines it, and fugitive doesn't know the git directory
properly at this point.

Setting `b:git_dir` to the temp .git directory fixes it for now, but
there is probably a better way of doing this.