Adjust gitlab_domains doc and support more cases
Issue #49 showed a mismatch between the docs the full SSH url.

We can both update the docs and catch more cases.

The issue with #49 was that we weren't allowing the user@ part of the
domain.
Whilst at the same time, the docs suggested having ssh:// in the
options when really we just need the domain with no schema, port or
user.

This adjusts both that use case and the README